Add controller error handling

+++ backend/controllers/productController.js
@@ -51,12 +51,15 @@ export const updateProduct = async (req,res) => {
 
 export const deleteProduct = async (req,res) => {
     const {id} = req.params;
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({success:false,message: "Product Not Found"});
+    }
 
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({ success:true, message: "Product Deleted"});
     } catch (error) {
         console.log("Error in deleting product:",error.message);
-        res.status(404).json({success:false,message:"Product not found"});
+        res.status(500).json({success:false,message:"Server Error"});
     }
 };
